one commit

+++ src/main/java/App/controller/UserManagement.java
@@ -1,13 +1,10 @@
 package App.controller;
 
-import java.io.IOException;
-import java.sql.SQLException;
-import java.util.Optional;
-
 import App.dataModel.UserData;
 import App.database.UserDb;
-import App.utile.MyDialog;
+import App.utile.Constant;
 import App.utile.FxmlUtile;
+import App.utile.MyDialog;
 import App.utile.ProgressFrom;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
@@ -22,53 +19,51 @@ import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.image.Image;
 import javafx.stage.Stage;
 
+import java.io.IOException;
+import java.sql.SQLException;
+import java.util.Optional;
+
 public class UserManagement {
 
+    Stage stage = new Stage();
     @FXML
     private TextField userQueryTextField;
-
     @FXML
     private TextField userNameTextField;
-
     @FXML
     private TextField jobNumTextField;
-
     @FXML
     private ComboBox<String> positionComboBox;
-
     @FXML
     private ComboBox<String> roleComboBox;
-
     @FXML
     private TextField telTextField;
-
     @FXML
     private TableView<UserData> personTable;
-
     @FXML
     private TableColumn<UserData, String> nameColumn;
-
     @FXML
     private TableColumn<UserData, String> telColumn;
-
     @FXML
     private TableColumn<UserData, String> jobNumColumn;
-
     @FXML
     private TableColumn<UserData, String> positionColumn;
-
     @FXML
     private TableColumn<UserData, String> roleColumn;
+    private ObservableList<String> positionList = FXCollections.observableArrayList("设计人员", "主管", "主任", "部长");
+    private ObservableList<String> roleList = FXCollections.observableArrayList("用户", "管理员", "超级管理员");
 
     @FXML
     void initialize() {
-        positionComboBox.setItems(positionList);
-        roleComboBox.setItems(roleList);
+        positionComboBox.setItems(FXCollections.observableArrayList(Constant.getPositionList()));
+        roleComboBox.setItems(FXCollections.observableArrayList(Constant.roleList));
+
         jobNumColumn.setCellValueFactory(new PropertyValueFactory<>("jobNum"));
         nameColumn.setCellValueFactory(new PropertyValueFactory<>("name"));
         telColumn.setCellValueFactory(new PropertyValueFactory<>("tel"));
         positionColumn.setCellValueFactory(new PropertyValueFactory<>("position"));
         roleColumn.setCellValueFactory(new PropertyValueFactory<>("role"));
+        personTable.setItems(FXCollections.observableArrayList(UserDb.getAllUser()));
         personTable.getSelectionModel().selectedItemProperty().addListener((observable, oldValue, newValue) -> showUserDetail(newValue));
     }
 
@@ -87,8 +82,12 @@ public class UserManagement {
     private void handleDeletePerson(ActionEvent event) throws IOException {
         UserData deletedUserData = personTable.getSelectionModel().getSelectedItem();
         String deletedUserJobNum = deletedUserData.getJobNum();
-        Optional<ButtonType> result = MyDialog.confirmation(null, "确认删除"+deletedUserData.getName()+"吗？");
+        Optional<ButtonType> result = MyDialog.confirmation(null, "确认删除" + deletedUserData.getName() + "吗？");
         if (result.get() == ButtonType.OK) {
+            if (UserDb.getSuperAdminNum() == 1 && deletedUserData.getRole().equals("超级管理员")) {
+                MyDialog.warning("您正在删除系统中唯一的超级管理员账号", "系统必须保证至少有一个超级管理员账号，因此该操作不能进行");
+                return;
+            }
             UserDb.delete(deletedUserJobNum);
             refreshPersonTable();
         }
@@ -103,7 +102,7 @@ public class UserManagement {
     }
 
     @FXML
-    private void handleEditPerson(ActionEvent event) throws IOException{
+    private void handleEditPerson(ActionEvent event) throws IOException {
         UserData editUserData = personTable.getSelectionModel().getSelectedItem();
         AddEditPerson addEditPerson = new AddEditPerson();
         addEditPerson.showAddEditPerson(editUserData);
@@ -133,14 +132,9 @@ public class UserManagement {
         }
     }
 
-    Stage stage = new Stage();
-
-    private ObservableList<String> positionList = FXCollections.observableArrayList("设计人员","主管","主任","部长");
-    private ObservableList<String> roleList = FXCollections.observableArrayList("用户","管理员","超级管理员");
-
     //刷新personTable的显示
     private void refreshPersonTable() {
-        personTable.setItems(UserDb.getUserDataList());
+        personTable.setItems(FXCollections.observableArrayList(UserDb.getAllUser()));
     }
 
     //显示用户管理界面
@@ -169,7 +163,7 @@ public class UserManagement {
 
     //关闭用户管理界面
     private void closeUserManagement(ActionEvent event) throws IOException {
-        ((Node)(event.getSource())).getScene().getWindow().hide();
+        ((Node) (event.getSource())).getScene().getWindow().hide();
     }
 
 }
